refactor(button): removes v2 support from button
