+++ packages/palette/src/elements/Button/index.ts
@@ -1,4 +1,4 @@
 export * from "./Button"
 export * from "./types"
-export { buttonMixin } from "./v3/Button"
-export * from "./v3/tokens"
+export { buttonMixin } from "./Button"
+export * from "./tokens"
