+++ packages/palette/src/elements/EntityHeader/EntityHeader.story.tsx
@@ -21,7 +21,7 @@ const FollowButton = () => {
     v3: (
       <Button
         size="small"
-        variant="secondaryOutline"
+        variant="secondaryBlack"
         onClick={() => alert("Follow")}
       >
         Follow
